updates to anova analysis
speed yaw curves made in two different ways, need to decide which is best ... one that is used for the anova doesn't balance for different number of data per mouse, because all "raw" data is pulled together. The older speed-yaw plots plot the means of the means per mouse. This is likely better, even though there will be much fewer data going in the ANOVA